requirements.txt
Django==3.2.9
django-cleanup==5.2.0
django-environ==0.8.1
django-summernote==0.8.20.0
Pillow==8.4.0
psycopg2==2.9.2
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1